Fix issue with deleted source resources
Reduce processor did not filtered deleted map resources, and map processor did
not marked deleted resources properly so in the end, everything seemed to work,
but in fact it didn't. At least for case, when reduce processor would be
interrupted, then it would not be able to resume correctly.

For that specific case, there is no test yet.